requirements.txt
anyio==3.7.0
click==8.1.3
colorama==0.4.6
fastapi==0.99.1
h11==0.14.0
idna==3.4
joblib==1.3.1
nltk==3.8.1
numpy==1.25.0
pandas==2.0.3
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.22.0
openpyxl==3.0.9
requests==2.26.0
